consolidate proxy web-tools tests into one integration file, add guard-bypass regression tests, and add a browser-e2e we… (W7)
